Update BiblioCraft to 1.11.7
BiblioCraft: update to BiblioCraft[v1.11.7][MC1.7.10].jar from
BiblioCraft[v1.11.5][MC1.7.10].jar

http://www.bibliocraftmod.com/2017/05/22/bibliocraft-updated-for-minecraft-v1-7-10/
- I just released an update to BiblioCraft, making it version 1.11.7 now
for Minecraft v1.7.10. This update fixes a crash that occurs when you
try to start a deticated server with BiblioCraft v1.11.6. There was a
bug I introduced on accident in the last update when I fixed a dupe with
the slotted book. I forgot make the call to open the gui happen only on
the client and let it happen on the server where it should never happy.
That is fixed now in this update.

http://www.bibliocraftmod.com/2017/05/19/bibliocraft-updated-for-minecraft-v1-11-2-and-many-important-bug-fixes-for-versions-for-minecraft-1-10-2-1-9-4-and-1-7-10/
- I also went through github and fixed many bugs and dupe issues that
have come up, including some issues that go back to the version for
1.7.10. So I went ahead and fixed a bunch of bugs in the version for
1.7.10, which is now updated to BiblioCraft v1.11.6.